Rename fieldparser to parser
