+++ plugins/processors/parser/parser_test.go
@@ -1,4 +1,4 @@
-package fieldparser
+package parser
 
 import (
 	"reflect"
